Get the first version going
- Config updates
- Initial drafts of the first post (missing images/proper highlighting)

TODO:
- About page
- Background image/avatar
- Post images
- Proper highlighting and formatting check

@DarylRodrigo Would be cool to get this out in the next few days. Commit
to this branch, then we can rebase it down and merge it in. Figured it
would be nice to work on a branch so that the production version wont
update till we're done.

To make it work locally:
Make sure 'url:' is blank in _config and then just run `bundle` followed
by `jekyll serve`